Remove early validation check in directives
The previously implemented early validation calls are in fact not necessary and have been removed:

If a symbol that is not defined yet is passed as a parameter, then it will pass the "potentiallyMatches" logic of the parameter checker.
However, even though the value is accessed during assignAddress (before validate() is called!), this is not a problem.

It does not matter if the symbol receives an incompatible type later, as it does not wait for the symbol to be resolved, since it performs its "work" during assignAddress(). Thus, instead of throwing an internal error due to an incompatible type, it will report a "symbol not found" error.

If the symbol is given an incompatible type before this directive, then it will not pass the parameter type checker.